+++ cmd/management-console/build.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 
+path_to_package=${1:-github.com/sourcegraph/sourcegraph/cmd/management-console}
 # We want to build multiple go binaries, so we use a custom build step on CI.
 cd $(dirname "${BASH_SOURCE[0]}")/../..
 set -ex
@@ -16,7 +17,7 @@ export GOARCH=amd64
 export GOOS=linux
 export CGO_ENABLED=0
 
-for pkg in github.com/sourcegraph/sourcegraph/cmd/management-console; do
+for pkg in $path_to_package; do
     go build -ldflags "-X github.com/sourcegraph/sourcegraph/pkg/version.version=$VERSION" -buildmode exe -tags dist -o $OUTPUT/$(basename $pkg) $pkg
 done
