+++ cmd/management-console/shared/auth_test.go
@@ -1,4 +1,4 @@
-package main
+package shared
 
 import (
 	"context"
