+++ cmd/management-console/shared/license.go
@@ -1,4 +1,4 @@
-package main
+package shared
 
 import (
 	"strconv"
