+++ enterprise/cmd/management-console/pre-build.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+cd $(dirname "${BASH_SOURCE[0]}")/../../..
+set -ex
+
+./cmd/management-console/pre-build.sh
